Subset polygons

Subset polygons from a sf or sfc objects.

st_subset_polygons(x)

Arguments

x
sf or sfc object.

Value

object of same class as argument to x.

Details

This function will remove any geometries that are not polygon or multipolygon geometries. Additionally, it will extract any polygon or multipolygon geometries from geometry collections.

Examples

# make data containing points, lines, and a geometry collection
## point
po <- sf::st_point(c(1, 2))

## line
lin <- sf::st_linestring(matrix(1:8, , 2))

## polygon
outer <- matrix(c(0, 0, 8, 0, 8, 8, 0, 8, 0, 0), ncol = 2, byrow = TRUE)
hole1 <- matrix(c(1, 1, 1, 2, 2, 2, 2, 1, 1, 1), ncol = 2, byrow = TRUE)
hole2 <- matrix(c(5, 5, 5, 6, 6, 6, 6, 5, 5, 5), ncol = 2, byrow = TRUE)
pl1 <- sf::st_polygon(list(outer))
pl2 <- sf::st_polygon(list(outer + 20, hole1 + 20, hole2 + 20))

## multi-polygon
mpl <- sf::st_multipolygon(list(list(outer + 30, hole1 + 30, hole2 + 30),
                                list(outer + 40, hole1 + 40),
                                list(outer + 50)))

## geometry collections
geomcol1 <- sf::st_geometrycollection(list(pl1 + 60, po + 60, pl2 + 60,
                                           lin + 60, mpl + 60))
geomcol2 <- sf::st_geometrycollection(list(po + 70, lin + 70))
geomcol3 <- sf::st_geometrycollection(list(pl1 + 80))
geomcol4 <- sf::st_geometrycollection()

# create sf object
x <- sf::st_sf(label = letters[1:9],
               geometry = sf::st_sfc(po, pl1, lin, pl2, mpl, geomcol1,
                                     geomcol2, geomcol3, geomcol4))

# subset polygons
y <- st_subset_polygons(x)

# print the data for visual comparisons
print(x)

#> Simple feature collection with 9 features and 1 field (with 1 geometry empty)
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 118 ymax: 118
#> epsg (SRID):    NA
#> proj4string:    NA
#>   label                       geometry
#> 1     a                    POINT (1 2)
#> 2     b POLYGON ((0 0, 8 0, 8 8, 0 ...
#> 3     c LINESTRING (1 5, 2 6, 3 7, ...
#> 4     d POLYGON ((20 20, 28 20, 28 ...
#> 5     e MULTIPOLYGON (((30 30, 38 3...
#> 6     f GEOMETRYCOLLECTION (POLYGON...
#> 7     g GEOMETRYCOLLECTION (POINT (...
#> 8     h GEOMETRYCOLLECTION (POLYGON...
#> 9     i       GEOMETRYCOLLECTION EMPTY

print(y)

#> Simple feature collection with 5 features and 1 field
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 118 ymax: 118
#> epsg (SRID):    NA
#> proj4string:    NA
#>   label                       geometry
#> 2     b POLYGON ((0 0, 8 0, 8 8, 0 ...
#> 4     d POLYGON ((20 20, 28 20, 28 ...
#> 5     e MULTIPOLYGON (((30 30, 38 3...
#> 6     f MULTIPOLYGON (((60 60, 68 6...
#> 8     h POLYGON ((80 80, 88 80, 88 ...