wdpar: Interface to the World Database on Protected Areas

The wdpar R package provides an interface to the World Database on Protected Areas (WDPA). It provides functions for automatically downloading data (from Protected Planet) and cleaning data following best practices (outlined in Butchart et al. 2015; Runge et al. 2015. The main functions are wdpa_fetch for downloading data and wdpa_clean for cleaning data. These functions are designed to be scalable and can employ parallel processing to reduce run time. For more information, please see the package vignette.